+++ docs/source/changelog.rst
@@ -28,6 +28,7 @@ Version 2.0
 * Make continuation tokens automatic and remove follow_continue parameter from API methods
 * Make auto_suggest default to False for page() and summary() methods
 * Fix deprecated BeautifulSoup findAll method
+* Add cache expiration mechanism (TTL) to memoization decorator
 
 Version 1.3
 =============